+++ lib/Cke5/Handler/Cke5DatabaseHandler.php
@@ -8,8 +8,6 @@
 namespace Cke5\Handler;
 
 
-use Cke5\Creator\Cke5ProfilesCreator;
-use DateTime;
 use DateTimeInterface;
 use rex;
 use rex_extension;